Update CreateNPCs.md
I felt that the couple of lines describing the NPC dialogue was very vague. Although there is not a lot to say about it from the default screen, I felt it should be made clearer that there is a basic dialogue built into the NPC and a more advance one via the use of behaviour packs and commands.