Change UUID capital F to to lowercase
VS Code gives a warning when a capital F is provided and the recommended extensions are installed:

Uuid

This is a unique identifier for the module in the same format as the pack's UUID in the header. This should be different from the pack's UUID, and different for every module

String does not match the pattern of "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$".