Split example in minecraftComponent_spawn_entity.md
Because this component can spawn an item or an entity, and its properties differ in each case, the example should be split into two distinct examples. The current example does not show the proper use of "spawn_item_event" despite spawning an item, and shows "spawn_entity" despite not spawning an entity.

Made one example spawn a chicken, while the other example spawns an egg.